Make blue-container style more accessible
Move the new blue-container style to a more accessible part of the project. This is to make it usable in the Modal component. There is also the idea to place it in a global.scss or something similar so that I can place the Modal in the <body> rather than the highest app element.